+++ bson/json_util.py
@@ -244,7 +244,7 @@ class JSONOptions(CodecOptions):
        Accepts the optional parameter `json_mode`.
 
     .. versionchanged:: 4.0
-        Changed default value of `tz_aware` to False
+       Changed default value of `tz_aware` to False
     """
 
     def __new__(cls, strict_number_long=False,